Move jotz into renderer/ and encapsulate
