+++ src/index.html
@@ -8,6 +8,6 @@
   <div id="sidemenu" class="side-container"></div>
   <div id="undermenu" class="under-container"></div>
   <div id="noteview" class="note-container"></div>
-  <script src="build/main.js"></script>
+  <script src="../build/build.js"></script>
 </body>
 </html>
