--- src/renderer/main.js
@@ -1,6 +1,0 @@
-//Sets up React var globally
-var React = require('react');
-
-var sideMenu = require('./build/components/sideMenu.js');
-var underMenu = require('./build/components/underMenu.js');
-var noteView = require('./build/components/noteView.js');
